Refactor: Renombrar carpeta ui/use-case a api y actualizar imports en main.js
