--- api/app.js
@@ -1,1 +1,0 @@
-let hola = "hola"
