--- ui/app3.js
@@ -1,1 +1,0 @@
-let hola = "digamen que tal"
